masc: update packages
