+++ masc-proxy/src/config.ts
@@ -1,0 +1,11 @@
+export const CONFIG = {
+    STAGENET_PORT: 18083,
+    STAGENET_HOST: 'localhost',
+    MAINNET_PORT: 18082,
+    MAINNET_HOST: 'localhost',
+}
+
+export enum MASC_ENV {
+    STAGENET = 'STAGENET',
+    MAINNET  = 'MAINNET'
+}
