+++ masc-proxy/src/routes/monero.ts
@@ -1,6 +1,11 @@
 const express = require("express");
 const router = express.Router();
 import axios from 'axios';
+import { CONFIG, MASC_ENV } from '../config';
+
+const isMainnet: boolean = process.env.MASC_PROXY_ENV === MASC_ENV.MAINNET
+const HOST = isMainnet ? CONFIG.MAINNET_HOST : CONFIG.STAGENET_HOST;
+const PORT = isMainnet ? CONFIG.MAINNET_PORT : CONFIG.STAGENET_PORT;
 
 // @route GET /proxy/monero/test
 // @desc Tests monero rpc route
@@ -19,7 +24,7 @@ router.get("/test", (_:any, res:any) =>
 router.post("/balance", (req:any, res:any) => {
   console.log(req.body);
     axios
-      .post('http://localhost:38083/json_rpc', req.body)
+      .post(`http://${HOST}:${PORT}/json_rpc`, req.body)
       .then((balance) => {
         console.log(`DEBUG: ${JSON.stringify(balance.data)}`);
         res.json( balance.data )
